code/django_rest_api/requirements.txt
beautifulsoup4==4.8.1
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
Django==2.1.7
django-allauth==0.41.0
django-cors-headers==2.5.2
django-rest-auth==0.9.5
djangorestframework==3.9.2
entrypoints==0.3
flake8==3.7.8
flake8-polyfill==1.0.2
idna==2.8
jaconv==0.2.4
mccabe==0.6.1
numpy==1.17.2
oauthlib==3.1.0
pandas==0.25.1
pep8-naming==0.8.2
psycopg2==2.7.6
psycopg2-binary==2.7.7
pycodestyle==2.5.0
pyflakes==2.1.1
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3
requests==2.22.0
requests-oauthlib==1.3.0
selenium==3.141.0
six==1.12.0
soupsieve==1.9.4
urllib3==1.25.6
